requirements.txt
Django>=2.1.4,<=2.2.0
djangorestframework>=3.9.0,<3.10.0\
markdown>=3.3.4       # Markdown support for the browsable API.
django-filter>=21.1